move image funcs to imgen, add place image function
